REFACT: Introduces test resource: DateTestDataBuilder
